much cleaner! since output depends only on currstate, and currstate is only ever updated on posedge clock, then output will only ever update on posedge clock, when currstate updates too. but, it would be dangerous to have output strictly update only on posedge clock --- it would be dangerous to have output be CLOCKED --- because then, behaviour of output depends on whether currstate was updated before or after output "code" is reached, creating a race condition.
by having output be a combinational circuit, continuously
computing result WITHOUT CLOCK, but the wires it depends on are
WITH CLOCK, output module behaves like it IS CLOCKED, but really
it is not.

currstate updates on posedge clock. since output will update
whenever its input currstate changes, it will update AFTER currstate
updates on posedge clock. there is nos magic here. output actually
CONTINUOUSLY AND INFINITELY updates itself given currstate. and if
currstate updates, then output will update too.

output is not clocked, but behaves as if it is clocked.
like how combinationalcircuit is not clocked, but also
behaves as if it were clocked (though, in our case it
clearly isn't clocked as it also depends on input w, which
is asynchronously set by the user flipping up and down
input w switch on fpga).