this should work well on fpga i hope it works well on fpga...
SW[9:2] are the instruction input. on transition from fetch to
decode, datapath will read from this instruction input, and save
its value in ISR as the next instruction to decode and execute.

SW[9:6] also control what is displayed on HEX's. 4'b0000 to 4'b0111
clearly would show the value of each of the eight registers.